requirements.txt
streamlit
streamlit_chat
groq
python-dotenv